Reformatting and fix the event state handlers
Since getChannelState and saveChannelState are both async with
callbacks, instead, we call getChannelState once on handle, and then
pass that value to each State (and StateEventHandler) that can respond
to the events.

Each one takes the channelState, may modifiy it, returns it, and passes
the new channelState to the next handler. Eventually, the final
channelState is saved out.

Also some code formatting. And got rid of logFunc, since I figured out
how to debug this with VS Code